box: introduce Log
The logger is just a couple of functions. The main one is declared
but not defined. The projects, which will use mgbox lib, should
define this function themselves.

This kind of link-time virtual function allows not to introduce
any global function pointers which people would have to setup
before using mgbox. The current approach allows to: 1) see all
needed functions-to-define during exe link-time, 2) not to pay for
virtual function calls.

The library mgboxstub implements the missing functions in some
trivial way to get the tests running, but shouldn't be used in
prod.