box: introduce TimePoint, TimeDuration, TimeLimit
Those allow to enforce usage of a time limit of a certain type.
This will be important in the network code where it is almost
always preferable to use deadlines in the internals; at the same
time the users usually give timeouts; and on the other hand - some
users might give already known deadlines to improve perf.