net: introduce DomainToIP
TODO:
- Extract thread.cpp change

The new module is able to asynchronously translate domain names
into IPs. This requires potential blocking work with the network,
so the resolution happens in a separate thread.

On Windows apparently it could be possible to use OVERLAPPED
version of getaddrinfo function to perform a truly async lookup.
But on Unix-systems the only alternative is getaddrinfo_a() which
internally is a thread pool anyway. Not kernel-level async. Better
have an own implementation then, which can be controlled.

This will be needed by the future network code.