[WIP] sio: introduce non-async networking
SIO stands for "synchronous IO". The module contains networking
code which behaves as simple as "do thing now" or "return error
and retry later". Basically, this API is intended for polling and
only for single-threaded usage.

It is not supposed to be used for anything non-trivial and is
mostly here for testing asynchronous IO coming in next commit, and
more networking things in the future.