test: rework TCPSocketIFace tests
The problem was that it was very hard to extend them at this
point. Each time when any work had to be done inside any of the
callbacks of TCPSocketSubscription, it was added to
TestClientSocket as new methods and members executed in those
callbacks. The test client was too clogged with things related to
many small test cases.

The patch reworks it in a way that now the test cases can quickly
subscribe to TCPSocketIFace events without inheritance, using
just simple C++ lambdas. It reminds a bit of boost::asio, but is
not suitable for the prod code due to performance overhead of all
this sugar. However for tests it is just fine. Allows to keep
TestClientSocket more independent from individual test cases.