aio: use release-asserts in all not hot places
The internals of aio might be a bit untrivial and it could be
easy to misuse in the first attempts for a new user. The misusage
is easier to detect when specific asserts are failing and not
just random crashes happen or UB.

The patch makes a lot of aio assertions active in the release
build too. The only debug-only ones now are those which are on
the perf-critical paths.