aio: fix empty buffer sending for TCPSocket
TCPSocket used to crash when got empty send buffers. Could happen
when there was enough of them to fill IOVec max count (128) and
then the underlying send system call would be invoked with 0 bytes
which leads to undefined behaviour in the system and to an
assertion in IOCore.

Lets skip empty buffers before trying to send anything. Trying to
filter them out at the stage of posting into the send queue is
possible, but it is too cumbersome to keep track of all places
where it can happen.