box: fix Buffer list to IOVec array converters
The used to take empty buffers too, assuming that it would cause
no harm.

It does indeed no harm in general when at least one buffer is not
empty, but better skip the empty ones anyway. The reason is that
during fresh benchmarks it was discovered that it is handy to be
able to push empty buffers into TCPSocketIFace, without having to
check whether they are empty not (can easily happen when read a
stream of requests, paste responses into a single buffer, and then
send it - when no requests, it will be empty).

That in turn means that the socket might get some empty buffers
slipped into it. If they would take space in the array of iovecs,
it would mean less of the non-empty ones per system call. Which is
bad for perf, potentially.

These if-checks cost nothing compared to win of making less system
calls, even in this unlikely scenario described above.