bench: allow to set scenario params per version
Previously it was only possible to give special params for the entire version for all its
scenarios. However for some scenarios it is nice to be able to provide version-specific
params just for these scenarios.

For example, there is going to be done a huge benchmark with 50 threads, which works fast
on the original TaskScheduler, but x100 slower on a trivial scheduler. It takes eternity
if they both run on the same task count. It is wanted to be able to give less tasks to the
trivial scheduler so its single run wouldn't take tens of minutes.