bench: 50 threads bench scenario for TaskScheduler
I got access to a really big machine with 32 physical cores and 64 logical ones. Had a
chance to try and bench TaskScheduler on it with lots of threads.

The results are impressive, it is up to x120 faster than a trivial scheduler on such
setup. And it fully utilizes the cores - they are all in 99% CPU during the run.