build: fix circular build dependency
mgboxstub depended on mgbox. But it shouldn't. Because box also implicitly
depends on symbols defined in the stub. It could result in incorrect order of
libs in linking process, which would leave the stub symbols unresolved.

Lets make the stub not depend on the lib which it is supposed to be patching.
Also add the stub to the benches as otherwise on some platforms they seem to try
and link with the logging symbols which fail unresolved without a stub.