aio: introduce SSLSocket
The socket is a new child-class of TCPSocketIFace. It implements
what one would expect from an SSL socket - looks like a TCP
stream, but is encrypted.

The public API is pretty much the same as for TCPSocket. Except
that during socket open it is possible to specify SSL-specific
parameters like an SSL context with all the settings; a hostname
(a TLS extension for SNI - Server Name Identification); and
whether the encryption must be enabled right away.

The optional encryption allows the user to start the SSL socket as
a regular TCP stream and only then hide it into an SSL tunnel.

This feature might be handy if the server doesn't want to spend
time on an expensive SSL handshake until the client does something
expensive too, as a DDoS protection. Some kind of Proof Of Work
which requires little effort on the server, but lots of work on
the client.

Closes #14