aio: preserve recv-queue between recv events
TCPSocket::myRecvQueue was cleared if during OnRecv() no new data
was requested.

It could happen that the caller simply needed more data for a
complete message and didn't consume the whole myRecvQueue.
Discarding this data simply leads to the message corruption.

Lets better preserve the data.

It is easy tested in this patch by trying to send large messages,
which won't be delivered in one piece.