aio: fix IOCore::Stop() hanging
During shutdown it could happen that the last IOCore worker would
be stuck inside the scheduler role, waiting on an infinite timeout
for any event.

The scenario was that when worker count > 1, the shutdown could
successfully stop all the workers besides the scheduler, because
it would simply by last in the shutdown chain. Then the scheduler
in the end was supposed to be woken up by the last
core.PrivPlatformSignal(). And it was indeed, but then it saw in
the end of IOCore::PrivScheduleDo(), that there are no events,
and it would enter an infinite sleep on poll().

The solution is to introduce a state for IOCore, which the
scheduler (IOCore::PrivScheduleDo()) could check and not enter the
infinite waiting then.

Closes #23