net: TCP backlog as a runtime value
On Apple the backlog is a runtime value, not a constant SOMAXCONN.
The constant exists, but might not reflect the reality when the
system value kern.ipc.somaxconn is changed.

Lets fetch this value at runtime, to match whatever is set in the
system. It is going to be useful for running more stressful tests
on Mac.