test: introduce test timeouts
Wait() template function can be used to wait for a condition or a
deadline expiration. When the deadline is over, the function
fails the tests.

It might be useful in CI where when a test is killed by CI itself,
the stdout doesn't seem to be available. But if the test gets
terminated on its own, the stdout buffer is supposed to be
flushed.

Also as an additional measure, the test prints are now all
followed by fflush(stdout);