test: fix aio stress reconnect
The test could hang on any system during termination.

In the end it disables the reconnect, posts a closure into each
client, and waits for its termination.

The problem is that it might happen that PostClose() (inside
CloseBlocking()) might happen when the socket is already closed
or closing, but still is inside of the reconnect-callback.

Then PostClose() is nop (it thinks the socket is already being
closed), and the reconnect does happen one last time,
successfully, and the WaitClose() hangs.

The solution is to drop all the reconnect callbacks before closing
their sockets.