aio: fix IOCore spurious wakeups on start
When IOCore is just started (IOCore::Start()), it would set its
state to RUNNING. But only after the workers are started. It
could lead to the workers spinning in busy-loops until the state
becomes RUNNING, because the workers would immediately exit the
scheduler when seeing STOPPED state.

Lets set the state RUNNING before the workers are launched. So as
they would properly wait for new events even before
IOCore::Start() is finished.