build: silence a warning about DoublyList
It would complain that the list iterator's member 'myNext'
sometimes isn't initialized.

It was supposed to be really. The case when it remains trash is
when the list is empty, and 'myPos' then is null. But then it is
equal to DoublyList::end() (which is also null) and the iteration
wouldn't even start.

Anyway, lets add the setting to just calm the compiler down.