test: fix mg::box::Coro use after free
It happened in the TaskScheduler tests. The reason was that some
of the coroutined tasks were allocated on the main thread's stack.

They would typically wait until their task sends a signal back to
the main thread, and then would destroy all the objects, including
the task itself.

The problem is that after the task in those tests would send the
signal, it would still have to do co_return. And that requires
the coroutine object to be still alive. Sometimes it could happen
that the main thread would already destroy the task and its
coroutine object by that time.

The solution is to send the signal only when the task is not going
to do any coroutine work anymore. In its callback's final suspend
code.

Closes #28