box: fix coroutines on Mac clang
There was a problem that if a coroutine operation inherited
mg::box::CoroOp and had its own members filled in its constructor,
then those members would be nullified despite there was no any
explicit change in the code to those members.

Strangely, it was only happening in GitHub CI macos-latest, and
only in one job (the other job in the same workflow was passing ok
even more tricky tests).

Eventually, debugging with prints right in CI has revealed that
apparently the compiler in `co_await SomeOperation()` copies this
operation. And if the operation has its copy/move constructors
deleted, then the members are not copied. In most cases simply
nullified.

That looks like a compiler bug, because the copying was explicitly
forbidden, and yet it happened. It was confirmed via prints -
operation's address (`this`) was one in constructor and another in
await_suspend().

Anyway, the fix seems to be just enabling the constructors.
Unfortunately, it is not really testable. All the coro tests
passed in macos CI and failed only in the future patch introducing
examples testing.