sch: don't receive front signal all the time
No need to receive it on each scheduling iteration. Under a high
load it won't change anything, because the queue is never empty.
The signal would be set again right away after receiving.

OTOH, not consuming it on each iteration doesn't break the logic,
because when the scheduler has nothing to do, it will sleep on the
signal. And if it was set, it would then be received.

This optimization leads to a spurious wakeup under no load. But
removes a needless signal receipt on a hot path.

The benchmarks though didn't show any difference. At least on
Apple Silicon M1. Anyway, this is now less code and the TLA+ model
is still correct and is slightly simpler.