sch: remove 'retry' from the scheduler
The idea previously was that a worker thread can't leave the
scheduler role until the scheduler is stopped or there are tasks
to do.

But actually nothing prevents from leaving the scheduler, entering
it again, and then sleeping on waiting for tasks.

It allows to simplify the state machine of the scheduler a bit.
Gives a guarantee that the scheduler-role will consume the front
signal only once per scheduling iteration.

This guarantee in turn is going to be quite handy in the future
potential API improvement for making it possible to run the
scheduler in user's threads.

Part of #40