box: introduce InterruptibleMutex
The new class is a special type of mutex which allows to take the
lock, go to sleep on some arbitrary condition (a condition
variable, a epoll_wait() call, a Signal.ReceiveBlocking(), or
something similar), and then get woken up when other threads want
to take the ownership.

This is an abstraction of the TaskScheduler's usecase. The
scheduler workers are going to be taking the lock and sleeping
when have to wait for new tasks or for a deadline. And other
threads would forcefully take the ownership when need to do some
maintenance work like stop/start of the scheduler or getting
something out of the scheduler-role context.

Currently there isn't much of the second type of work, but this
will change in scope of #40.

Part of #40